+++ mne/decoding/__init__.py
@@ -5,7 +5,7 @@ from .transformer import (PSDEstimator, Vectorizer,
                           UnsupervisedSpatialFilter, TemporalFilter)
 from .mixin import TransformerMixin
 from .base import BaseEstimator, LinearModel, get_coef, cross_val_multiscore
-from .csp import CSP
+from .csp import CSP, SPoC
 from .ems import compute_ems, EMS
 from .time_gen import GeneralizationAcrossTime, TimeDecoding
 from .time_frequency import TimeFrequency