+++ mne/decoding/csp.py
@@ -15,6 +15,7 @@ from scipy import linalg
 from .mixin import TransformerMixin
 from .base import BaseEstimator
 from ..cov import _regularized_covariance
+from ..utils import warn
 
 
 class CSP(TransformerMixin, BaseEstimator):
@@ -49,34 +50,44 @@ class CSP(TransformerMixin, BaseEstimator):
         If 'average_power' then self.transform will return the average power of
         each spatial filter. If 'csp_space' self.transform will return the data
         in CSP space. Defaults to 'average_power'.
+    norm_trace : bool
+        Normalize class covariance by its trace. Defaults to True. Trace
+        normalization is a step of the original CSP algorithm [1]_ to eliminate
+        magnitude variations in the EEG between individuals. It is not applied
+        in more recent work [2]_, [3]_ and can have a negative impact on
+        patterns ordering.
 
     Attributes
     ----------
-    ``filters_`` : ndarray, shape (n_channels, n_channels)
+    ``filters_`` : ndarray, shape (n_components, n_channels)
         If fit, the CSP components used to decompose the data, else None.
-    ``patterns_`` : ndarray, shape (n_channels, n_channels)
+    ``patterns_`` : ndarray, shape (n_components, n_channels)
         If fit, the CSP patterns used to restore M/EEG signals, else None.
     ``mean_`` : ndarray, shape (n_components,)
         If fit, the mean squared power for each component.
     ``std_`` : ndarray, shape (n_components,)
         If fit, the std squared power for each component.
 
+    See Also
+    --------
+    mne.preprocessing.Xdawn, SPoC
+
     References
     ----------
-    [1] Zoltan J. Koles, Michael S. Lazar, Steven Z. Zhou. Spatial Patterns
-        Underlying Population Differences in the Background EEG. Brain
-        Topography 2(4), 275-284, 1990.
-    [2] Benjamin Blankertz, Ryota Tomioka, Steven Lemm, Motoaki Kawanabe,
-        Klaus-Robert Müller. Optimizing Spatial Filters for Robust EEG
-        Single-Trial Analysis. IEEE Signal Processing Magazine 25(1), 41-56,
-        2008.
-    [3] Grosse-Wentrup, Moritz, and Martin Buss. Multiclass common spatial
-        patterns and information theoretic feature extraction. IEEE
-        Transactions on Biomedical Engineering, Vol 55, no. 8, 2008.
+    .. [1] Zoltan J. Koles, Michael S. Lazar, Steven Z. Zhou. Spatial Patterns
+           Underlying Population Differences in the Background EEG. Brain
+           Topography 2(4), 275-284, 1990.
+    .. [2] Benjamin Blankertz, Ryota Tomioka, Steven Lemm, Motoaki Kawanabe,
+           Klaus-Robert Müller. Optimizing Spatial Filters for Robust EEG
+           Single-Trial Analysis. IEEE Signal Processing Magazine 25(1), 41-56,
+           2008.
+    .. [3] Grosse-Wentrup, Moritz, and Martin Buss. Multiclass common spatial
+           patterns and information theoretic feature extraction. IEEE
+           Transactions on Biomedical Engineering, Vol 55, no. 8, 2008.
     """
 
     def __init__(self, n_components=4, reg=None, log=None, cov_est="concat",
-                 transform_into='average_power'):
+                 transform_into='average_power', norm_trace=None):
         """Init of CSP."""
         # Init default CSP
         if not isinstance(n_components, int):
@@ -115,6 +126,15 @@ class CSP(TransformerMixin, BaseEstimator):
                 raise ValueError('log must be a None if transform_into == '
                                  '"csp_space".')
         self.log = log
+        if norm_trace is None:
+            norm_trace = True
+            warn("new_param defaults to True in 0.15, but will change to False"
+                 " in 0.16. Set it explicitly to avoid this warning.",
+                 DeprecationWarning)
+
+        if not isinstance(norm_trace, bool):
+            raise ValueError('norm_trace must be a bool.')
+        self.norm_trace = norm_trace
 
     def _check_Xy(self, X, y=None):
         """Aux. function to check input data."""
@@ -166,8 +186,17 @@ class CSP(TransformerMixin, BaseEstimator):
                 cov /= len(class_)
                 weight = len(class_)
 
-            # normalize by trace and stack
-            covs[class_idx] = cov / np.trace(cov)
+            covs[class_idx] = cov
+            if self.norm_trace:
+                # Append covariance matrix and weight. Prior to version 0.15,
+                # trace normalization was applied, but was breaking results for
+                # some usecases by chaging the apparent ranking of patterns.
+                # Trace normalization of the covariance matrix was removed
+                # without signigificant effect on patterns or performances.
+                # If the user interested in this feature, we suggest trace
+                # normalization of the epochs prior to the CSP.
+                covs[class_idx] /= np.trace(cov)
+
             sample_weights.append(weight)
 
         if n_classes == 2:
@@ -215,7 +244,7 @@ class CSP(TransformerMixin, BaseEstimator):
         X = np.asarray([np.dot(pick_filters, epoch) for epoch in X])
 
         # compute features (mean band power)
-        X = (X ** 2).mean(axis=-1)
+        X = (X ** 2).mean(axis=2)
 
         # To standardize features
         self.mean_ = X.mean(axis=0)
@@ -250,7 +279,7 @@ class CSP(TransformerMixin, BaseEstimator):
 
         # compute features (mean band power)
         if self.transform_into == 'average_power':
-            X = (X ** 2).mean(axis=-1)
+            X = (X ** 2).mean(axis=2)
             log = True if self.log is None else self.log
             if log:
                 X = np.log(X)
@@ -267,18 +296,18 @@ class CSP(TransformerMixin, BaseEstimator):
                       show_names=False, title=None, mask=None,
                       mask_params=None, outlines='head', contours=6,
                       image_interp='bilinear', average=None, head_pos=None):
-        """Plot topographic patterns of CSP components.
+        """Plot topographic patterns of components.
 
-        The CSP patterns explain how the measured data was generated
-        from the neural sources (a.k.a. the forward model).
+        The patterns explain how the measured data was generated from the
+        neural sources (a.k.a. the forward model).
 
         Parameters
         ----------
         info : instance of Info
-            Info dictionary of the epochs used to fit CSP.
+            Info dictionary of the epochs used for fitting.
             If not possible, consider using ``create_info``.
         components : float | array of floats | None.
-           The CSP patterns to plot. If None, n_components will be shown.
+           The patterns to plot. If None, n_components will be shown.
         ch_type : 'mag' | 'grad' | 'planar1' | 'planar2' | 'eeg' | None
             The channel type to plot. For 'grad', the gradiometers are
             collected in pairs and the RMS for each pair is plotted.
@@ -422,18 +451,18 @@ class CSP(TransformerMixin, BaseEstimator):
                      show_names=False, title=None, mask=None,
                      mask_params=None, outlines='head', contours=6,
                      image_interp='bilinear', average=None, head_pos=None):
-        """Plot topographic filters of CSP components.
+        """Plot topographic filters of components.
 
-        The CSP filters are used to extract discriminant neural sources from
+        The filters are used to extract discriminant neural sources from
         the measured data (a.k.a. the backward model).
 
         Parameters
         ----------
         info : instance of Info
-            Info dictionary of the epochs used to fit CSP.
+            Info dictionary of the epochs used for fitting.
             If not possible, consider using ``create_info``.
         components : float | array of floats | None.
-           The CSP patterns to plot. If None, n_components will be shown.
+           The patterns to plot. If None, n_components will be shown.
         ch_type : 'mag' | 'grad' | 'planar1' | 'planar2' | 'eeg' | None
             The channel type to plot. For 'grad', the gradiometers are
             collected in pairs and the RMS for each pair is plotted.
@@ -593,9 +622,9 @@ def _ajd_pham(X, eps=1e-6, max_iter=15):
 
     References
     ----------
-    [1] Pham, Dinh Tuan. "Joint approximate diagonalization of positive
-    definite Hermitian matrices." SIAM Journal on Matrix Analysis and
-    Applications 22, no. 4 (2001): 1136-1152.
+    .. [1] Pham, Dinh Tuan. "Joint approximate diagonalization of positive
+           definite Hermitian matrices." SIAM Journal on Matrix Analysis and
+           Applications 22, no. 4 (2001): 1136-1152.
 
     """
     # Adapted from http://github.com/alexandrebarachant/pyRiemann
@@ -652,3 +681,154 @@ def _ajd_pham(X, eps=1e-6, max_iter=15):
             break
     D = np.reshape(A, (n_times, -1, n_times)).transpose(1, 0, 2)
     return V, D
+
+
+class SPoC(CSP):
+    """Implementation of the SPoC spatial filtering.
+
+    Source Power Comodulation (SPoC) [1]_ allows to extract spatial filters and
+    patterns by using a target (continuous) variable in the decomposition
+    process in order to give preference to components whose power correlates
+    with the target variable.
+
+    SPoC can be seen as an extension of the CSP driven by a continuous
+    variable rather than a discrete variable. Typical applications include
+    extraction of motor patterns using EMG power or audio patterns using sound
+    envelope.
+
+    Parameters
+    ----------
+    n_components : int
+        The number of components to decompose M/EEG signals.
+    reg : float | str | None, defaults to None
+        if not None, allow regularization for covariance estimation
+        if float, shrinkage covariance is used (0 <= shrinkage <= 1).
+        if str, optimal shrinkage using Ledoit-Wolf Shrinkage ('ledoit_wolf')
+        or Oracle Approximating Shrinkage ('oas').
+    log : None | bool, defaults to None
+        If transform_into == 'average_power' and log is None or True, then
+        applies a log transform to standardize the features, else the features
+        are z-scored. If transform_into == 'csp_space', then log must be None.
+    transform_into : {'average_power', 'csp_space'}
+        If 'average_power' then self.transform will return the average power of
+        each spatial filter. If 'csp_space' self.transform will return the data
+        in CSP space. Defaults to 'average_power'.
+
+    Attributes
+    ----------
+    ``filters_`` : ndarray, shape (n_components, n_channels)
+        If fit, the SPoC spatial filters, else None.
+    ``patterns_`` : ndarray, shape (n_components, n_channels)
+        If fit, the SPoC spatial patterns, else None.
+    ``mean_`` : ndarray, shape (n_components,)
+        If fit, the mean squared power for each component.
+    ``std_`` : ndarray, shape (n_components,)
+        If fit, the std squared power for each component.
+
+    See Also
+    --------
+    mne.preprocessing.Xdawn, CSP
+
+    References
+    ----------
+    .. [1] Dahne, S., Meinecke, F. C., Haufe, S., Hohne, J., Tangermann, M.,
+           Muller, K. R., & Nikulin, V. V. (2014). SPoC: a novel framework for
+           relating the amplitude of neuronal oscillations to behaviorally
+           relevant parameters. NeuroImage, 86, 111-122.
+    """
+
+    def __init__(self, n_components=4, reg=None, log=None,
+                 transform_into='average_power'):
+        """Init of SPoC."""
+        super(SPoC, self).__init__(n_components=n_components, reg=reg, log=log,
+                                   cov_est="epoch", norm_trace=False,
+                                   transform_into=transform_into)
+        # Covariance estimation have to be done on the single epoch level,
+        # unlike CSP where covariance estimation can also be achieved through
+        # concatenation of all epochs from the same class.
+        delattr(self, 'cov_est')
+        delattr(self, 'norm_trace')
+
+    def fit(self, X, y):
+        """Estimate the SPoC decomposition on epochs.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_epochs, n_channels, n_times)
+            The data on which to estimate the SPoC.
+        y : array, shape (n_epochs,)
+            The class for each epoch.
+
+        Returns
+        -------
+        self : instance of SPoC
+            Returns the modified instance.
+        """
+        if not isinstance(X, np.ndarray):
+            raise ValueError("X should be of type ndarray (got %s)."
+                             % type(X))
+        self._check_Xy(X, y)
+
+        if len(np.unique(y)) < 2:
+            raise ValueError("y must have at least two distinct values.")
+
+        # The following code is direclty copied from pyRiemann
+
+        # Normalize target variable
+        target = y.astype(np.float64)
+        target -= target.mean()
+        target /= target.std()
+
+        n_epochs, n_channels = X.shape[:2]
+
+        # Estimate single trial covariance
+        covs = np.empty((n_epochs, n_channels, n_channels))
+        for ii, epoch in enumerate(X):
+            covs[ii] = _regularized_covariance(epoch, reg=self.reg)
+
+        C = covs.mean(0)
+        Cz = np.mean(covs * target[:, np.newaxis, np.newaxis], axis=0)
+
+        # solve eigenvalue decomposition
+        evals, evecs = linalg.eigh(Cz, C)
+        evals = evals.real
+        evecs = evecs.real
+        # sort vectors
+        ix = np.argsort(np.abs(evals))[::-1]
+
+        # sort eigenvectors
+        evecs = evecs[:, ix].T
+
+        # spatial patterns
+        self.patterns_ = linalg.pinv(evecs).T  # n_channels x n_channels
+        self.filters_ = evecs  # n_channels x n_channels
+
+        pick_filters = self.filters_[:self.n_components]
+        X = np.asarray([np.dot(pick_filters, epoch) for epoch in X])
+
+        # compute features (mean band power)
+        X = (X ** 2).mean(axis=-1)
+
+        # To standardize features
+        self.mean_ = X.mean(axis=0)
+        self.std_ = X.std(axis=0)
+
+        return self
+
+    def transform(self, X):
+        """Estimate epochs sources given the SPoC filters.
+
+        Parameters
+        ----------
+        X : array, shape (n_epochs, n_channels, n_times)
+            The data.
+
+        Returns
+        -------
+        X : ndarray
+            If self.transform_into == 'average_power' then returns the power of
+            CSP features averaged over time and shape (n_epochs, n_sources)
+            If self.transform_into == 'csp_space' then returns the data in CSP
+            space and shape is (n_epochs, n_sources, n_times)
+        """
+        return super(SPoC, self).transform(X)